func_80016784 takes s32, s32 -- not s16, per a matched ported caller
The prototype said `s16 arg2, s16 arg3`, which makes the callee sign-extend
its own incoming arguments (`sll 16 / sra 16`) where retail does not.

This is not a guess from the listing: src/func_80015DFC.c is a file ported
from Unchiga's tree that builds byte-identical, and it declares the function
`(Inner *a0, s32 a1, s32 a2, s32 a3)`. Two trees agreeing on a signature is
better evidence than either alone.

Checked the way this repo requires when a signature changes: the affected
caller func_80016D04 still MATCHES, and the full build is still byte-identical
at 84747e64. src/func_80015DFC.c is self-contained and keeps its own
declarations, so it was never affected.

Worth 4 instructions on the func_80016784 draft in progress (+40 to +36).

+++ include/functions.h
@@ -47,7 +47,7 @@ void func_8004020C(s32 arg0);
 void func_80045334(s32 arg0);
 void func_80047314(s32 arg0);
 void func_80047430(s32 arg0, s32 arg1);
-void func_80016784(u8 *arg0, s32 arg1, s16 arg2, s16 arg3);
+void func_80016784(u8 *arg0, s32 arg1, s32 arg2, s32 arg3);
 /* Takes a second argument: func_80023FBC sets $a1 and func_80023D08 reads it
  * before writing it (bltz $a1). The one-argument form was a guess. */
 void func_80023D08(u8 *arg0, s32 arg1);
